complete fixing currency to mmk

+++ database/migrations/2023_05_08_025563_create_payments_table.php
@@ -16,7 +16,7 @@ class CreatePaymentsTable extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
             $table->string('order_code');
-            $table->decimal('payment_amount', 20, 2);
+            $table->integer('payment_amount');
             $table->string('payment_currency');
             $table->string('payment_method');
             $table->string('payment_gateway');
